+++ pages/about-us.js
@@ -1,9 +1,0 @@
-export default function AboutUs() {
-  return (
-    <div className="about-us">
-      <h1>About Us</h1>
-      <p>We're building this page.</p>
-      <p>Stay tuned!</p>
-    </div>
-  );
-}